requirements.txt
playwright>=1.40.0
mistralai>=1.0.0
openai>=1.0.0
PyPDF2>=3.0.0
python-dotenv>=1.0.0
python-pptx>=0.6.23
comtypes>=1.4.1
flask>=3.0.0
